Closes #22: Implementa botones "Aceptar" y "Cancelar"
Se implementaron los botones "Aceptar" y "Cancelar" para ser usado en
cada formulario faltante, se modifico el formulario de registro para las
solicitudes de escuelas.
