+++ lib/Persona/Colaborador/ui/screens/Responsive/colaboradores_desktop.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:suenomotora_app/Persona/Colaborador/ui/widgets/colaborador_card.dart';
 import 'package:suenomotora_app/Persona/Colaborador/ui/widgets/form_diaglog_detalles_colaboradores.dart';
+import 'package:suenomotora_app/Persona/Colaborador/ui/widgets/form_dialog_registro_colaboradores.dart';
 import 'package:suenomotora_app/common/widgets/floating_button.dart';
-import 'package:suenomotora_app/common/widgets/forms_dialogs.dart';
 
 class ColaboradoresDesktop extends StatelessWidget {
   const ColaboradoresDesktop({Key? key}) : super(key: key);
@@ -121,6 +121,6 @@ class ColaboradoresDesktop extends StatelessWidget {
         ),
         floatingActionButton: floButton.floatingButton(
             btnAction: () =>
-                FormsDialogs.mostrarDialogRegistroColaborador(context)));
+                FormDialogRegistroColaboradores.formDialogRegistroColaboradores(context)));
   }
 }
