+++ lib/Persona/Colaborador/ui/screens/Responsive/colaboradores_tablet.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:suenomotora_app/Persona/Colaborador/ui/widgets/colaborador_card.dart';
 import 'package:suenomotora_app/Persona/Colaborador/ui/widgets/form_diaglog_detalles_colaboradores.dart';
 import 'package:suenomotora_app/common/widgets/floating_button.dart';
-import 'package:suenomotora_app/common/widgets/forms_dialogs.dart';
+
+import '../../widgets/form_dialog_registro_colaboradores.dart';
 
 class ColaboradoresTablet extends StatelessWidget {
   const ColaboradoresTablet({Key? key}) : super(key: key);
@@ -48,6 +49,7 @@ class ColaboradoresTablet extends StatelessWidget {
         ),
         floatingActionButton: floButton.floatingButton(
             btnAction: () =>
-                FormsDialogs.mostrarDialogRegistroColaborador(context)));
+                FormDialogRegistroColaboradores.formDialogRegistroColaboradores(
+                    context)));
   }
 }
