+++ lib/Persona/Colaborador/ui/widgets/form_dialog_registro_colaboradores.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:suenomotora_app/common/widgets/forms_elements.dart';
 
-class FormsDialogs extends StatefulWidget {
-  const FormsDialogs({Key? key}) : super(key: key);
+class FormDialogRegistroColaboradores extends StatefulWidget {
+  const FormDialogRegistroColaboradores({Key? key}) : super(key: key);
 
-  static mostrarDialogRegistroColaborador(BuildContext context) {
+  static formDialogRegistroColaboradores(BuildContext context) {
     FormsElements formsElements = FormsElements();
 
     return showDialog(
