+++ lib/Solicitudes/ui/screens/responsive/solicitudes_escuelas_movil.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:suenomotora_app/Solicitudes/ui/widgets/form_dialog_registro_solicitudes_escuelas.dart';
 import 'package:suenomotora_app/Solicitudes/ui/widgets/forms_dialog_detalle_solicitudes_escuelas.dart';
 import 'package:suenomotora_app/common/widgets/floating_button.dart';
 import '../../widgets/card_solicitudes_escuelas.dart';
@@ -26,8 +27,9 @@ class SolicitudesEscuelasMovil extends StatelessWidget {
           ],
         ),
       ),
-      floatingActionButton:
-          floButton.floatingButton(btnAction: () => print('hola')),
+      floatingActionButton: floButton.floatingButton(
+          btnAction: () => FormDialogRegistroSolicitudEscuela
+              .formDialogRegistroSolicitudEscuela(context)),
     );
   }
 }
