+++ lib/common/widgets/forms_elements.dart
@@ -76,6 +76,24 @@ class FormsElements extends StatelessWidget {
     );
   }
 
+  Widget btnsAceptarCancelar(context) {
+    return Padding(
+      padding: const EdgeInsets.all(30.0),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.end,
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        children: [
+          ElevatedButton(
+              onPressed: () => Navigator.of(context).pop(),
+              child: const Text('Cancelar')),
+          ElevatedButton(
+              onPressed: () => Navigator.of(context).pop(),
+              child: const Text('Ok')),
+        ],
+      ),
+    );
+  }
+
   Widget crearEmail() {
     return Padding(
       padding: const EdgeInsets.all(8.0),
